How ignoring left Alt only if keyboard ID indicates that it is the default hardware keyboard (is 0).
This allows us to use it for symbol input on such keyboards and for Alt in any other scenario.

Additionally, a bugfix for aRDP and aSPICE where a static method needed to become public in aRDP.java
and aSPICE.java.

Finally, there is a bugfix for aSPICE where the native libraries were loaded in aSPICE and if an
intent is used to start RemoteCanvasActivity directly they would not be loaded. The loading of
native libs and initializing of Gstreamer was moved back to SpiceCommunicator.java.