Bugfix for Mac OS X when screen-sharing is done by password.
The addition of auth type 30 had broken backward compatibility with
Mac OS X machines configured to share screen if a client authenticates
via a VNC password. bVNC was picking auth type 30 even if the user name
field was not set in the settings, and when this auth type was chosen
Mac OS X was expecting a user name to be passed as well as the password
for that user name rather than a VNC password. Now, if a user name is not
supplied to bVNC, auth type 30 is not chosen and authentication falls
through to auth type 2.