Several changes to stunnel support to reduce complexity and improve security.
Now using validateX509Cert() for both X509Tunnel and SecureTunnel. That
function always saves the cert data upon user approval and then treats it as
definitive for future connections.

Moved the code responsible for setting up the secure tunnel to RfbProto to
be in line with how VeNCrypt and AnonTLS support is implemented.

No longer saving hashId and hashIdAlgorithm, as they are pertinent only for
the initial connection establishment.

Improved messages for related dialogs and translations to Russian are included.